PRKC-686-----
Added by Suneel Allareddy

There are issues with using int instead of long. 

soem scenorios:-
Form expiration of 35 days results in a due date 15 days PRIOR to the start date
Form expiration of 30 days results in a due date 20 days PRIOR to the start date
Form expiration of 25 days results in a due date 25 days PRIOR to the start date
Form expiration of 20 days results in a due date 1 day AFTER to the start date

+++ pro-ctcae/software/core/src/main/java/gov/nih/nci/ctcae/core/domain/ParticipantSchedule.java
@@ -92,7 +92,7 @@ public class ParticipantSchedule {
      */
     public void createSchedules(ScheduleType scheduleType) throws ParseException {
         proCtcAECalendar.prepareSchedules(scheduleType);
-        int dueAfterPeriodInMill = proCtcAECalendar.getDueAfterPeriodInMill();
+        long dueAfterPeriodInMill = proCtcAECalendar.getDueAfterPeriodInMill();
         while (proCtcAECalendar.hasMoreSchedules()) {
             if (scheduleType.equals(ScheduleType.GENERAL)) {
                 createSchedule(proCtcAECalendar.getNextGeneralScehdule(), dueAfterPeriodInMill, -1, -1, null, false);
