+++ pro-ctcae/software/web/src/main/java/gov/nih/nci/ctcae/web/participant/AddCrfScheduleController.java
@@ -40,7 +40,7 @@ public class AddCrfScheduleController extends AbstractController {
         ParticipantSchedule participantSchedule = studyParticipantCommand.getParticipantSchedules().get(index);
 
         Calendar c = new GregorianCalendar();
-        int duedate = 24 * 60 * 60 * 1000;
+        long duedate = 24 * 60 * 60 * 1000;
         if ("delall".equals(action)) {
             participantSchedule.removeAllSchedules(formIds);
         }
